parse: change parsing hard clauses
The definition of the MaxSAT competition format asks to label hard clauses
with the weight of the top value of the header line of the MaxSAT file.

As the top value can be seen as the maximum value that is considered as
solution, any soft clause that has a greater weight should never be
falsified, and hence could become a hard clause.

This commit turns clauses with this property (weight greater than top
value) into hard clauses immediately. This matches the competition format,
and should futhermore speedup open-wbo on formulas that contain soft
clauses whose weight is greater. While in the competition these formulas
should not be present, they might occur in the wild, and open-wbo should
be able to work fast on them.