model: print fast
While the original implementation runs one printf command for each
variable, which might be slow due to the involved IO operation, the new
implementation feeds the model to be printed into a string stream. This
way, the model is constructed into fast memory. Next, the whole model is
printed in a single printf, resulting in printing the model faster.